requirements.txt
flask
psycopg2-binary
redis
aioredis
pillow
pymupdf
fitz
google-generativeai
asyncpg
aiosqlite
Flask[async]